+++ presentation/src/main/java/com/ezlo/mydevices/presentation/FragmentCallback.kt
@@ -3,5 +3,5 @@ package com.ezlo.mydevices.presentation
 interface FragmentCallback {
     fun handleLoaderVisibility(isLoaderVisible: Boolean)
 
-    fun handleDeleteDevice(deviceSn: Long)
+    fun handleDeleteDevice(deviceSn: Int)
 }
